move functional test entities to separate file
