+++ src/test/kotlin/r2dbcfun/test/functional/RepositoryFunctionalTest.kt
@@ -29,23 +29,6 @@ import java.time.LocalDate
 private val characters = ('A'..'Z').toList() + (('a'..'z').toList()).plus(' ')
 private val reallyLongString = (1..20000).map { characters.random() }.joinToString("")
 
-data class UserPK(override val id: Long) : PK
-enum class Color {
-    RED,
-
-    @Suppress("unused")
-    BLUE
-}
-
-data class User(
-    val id: UserPK? = null,
-    val name: String,
-    val email: String?,
-    val isCool: Boolean? = false,
-    val bio: String? = null,
-    val favoriteColor: Color? = null,
-    val birthday: LocalDate? = null
-)
 
 class RepositoryFunctionalTest : FunSpec({
     forAllDatabases(this) { connection ->
