+++ src/test/kotlin/r2dbcfun/test/functional/User.kt
@@ -1,0 +1,22 @@
+package r2dbcfun.test.functional
+
+import r2dbcfun.PK
+import java.time.LocalDate
+
+data class UserPK(override val id: Long) : PK
+enum class Color {
+    RED,
+
+    @Suppress("unused")
+    BLUE
+}
+
+data class User(
+    val id: UserPK? = null,
+    val name: String,
+    val email: String?,
+    val isCool: Boolean? = false,
+    val bio: String? = null,
+    val favoriteColor: Color? = null,
+    val birthday: LocalDate? = null
+)
